changes of names, creating story class
